tools/vrrtest_tool: Add new tool to test vrr
Introduce a tool to show Flip Done events with or without vrr.
The tool provides an option to choose a flip rate and to enable/disable
vrr. It flips continously at the given rate, and logs flip and flip
done event timestamps.
It then computes the interval after which the flip done events are
received for the given mode and displays the interval between flips.

Signed-off-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>